refactor(editor): move hljs theme switching to code block component
