+++ packages/web-pkg/src/editor/components/CodeBlockComponent.vue
@@ -18,17 +18,26 @@
 </template>
 
 <script setup lang="ts">
-import { computed } from 'vue'
+import { computed, unref, watch } from 'vue'
 import { NodeViewContent, nodeViewProps, NodeViewWrapper } from '@tiptap/vue-3'
 import { useGettext } from 'vue3-gettext'
+import { storeToRefs } from 'pinia'
+import { useThemeStore } from '../../composables'
+import atomOneDarkThemeUrl from 'highlight.js/styles/atom-one-dark.css?url'
+import atomOneLightThemeUrl from 'highlight.js/styles/atom-one-light.css?url'
 
 const props = defineProps(nodeViewProps)
 const { $gettext } = useGettext()
+const themeStore = useThemeStore()
+const { currentTheme } = storeToRefs(themeStore)
 
 type LanguageOption = {
   label: string
   value: string | null
 }
+
+const isDarkTheme = computed(() => unref(currentTheme)?.isDark)
+const hljsThemeStylesheetId = 'oc-text-editor-hljs-theme'
 
 const languages = computed<string[]>(() => {
   return props.extension.options.lowlight.listLanguages() as string[]
@@ -58,4 +67,24 @@ const selectedLanguageOption = computed<LanguageOption>(() => {
 function updateSelectedLanguage(option: LanguageOption | null) {
   selectedLanguage.value = option?.value ?? null
 }
+
+watch(
+  isDarkTheme,
+  (isDark) => {
+    if (typeof document === 'undefined') {
+      return
+    }
+
+    let stylesheet = document.getElementById(hljsThemeStylesheetId) as HTMLLinkElement | null
+    if (!stylesheet) {
+      stylesheet = document.createElement('link')
+      stylesheet.id = hljsThemeStylesheetId
+      stylesheet.rel = 'stylesheet'
+      document.head.appendChild(stylesheet)
+    }
+
+    stylesheet.href = isDark ? atomOneDarkThemeUrl : atomOneLightThemeUrl
+  },
+  { immediate: true }
+)
 </script>
